+++ R/ggemmeans.R
@@ -166,8 +166,8 @@ ggemmeans <- function(model,
     "asymptotic"
   else if (inherits(model, "MixMod"))
     "fixed-effects"
-  else if (inherits(model, "gls"))
-    "satterthwaite"
+  else if (inherits(model, c("gls", "lme")))
+    "auto"
   else if (inherits(model, "MCMCglmm") && model_info$is_multinomial)
     "response"
   else if (model_info$is_ordinal || model_info$is_categorical || model_info$is_multinomial)